% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fdr.r
\name{fdp}
\alias{fdp}
\title{A implement of FDP and limit FDP at t}
\usage{
fdp(t)
}
\arguments{
\item{t}{the point to be estimated}
}
\value{
a list of FDP and limit FDP
}
\description{
A implement of FDP and limit FDP at t
}
\examples{
\dontrun{
set.seed(12345)
n <- 100; rho <- 0.5; sig <- 2; p.nonzero <- 10; beta.nonzero <- 1
fdp(0.01)
}
}
