% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fdr.r
\name{my.fdp}
\alias{my.fdp}
\title{A implement of FDP and limit FDP at t}
\usage{
my.fdp(t)
}
\arguments{
\item{t}{the point to be estimated}
}
\value{
a list of FDP and limit FDP
}
\description{
A implement of FDP and limit FDP at t
}
\examples{
\dontrun{
set.seed(12345)
n <- 100; rho <- 0.5; sig <- 2; p.nonzero <- 10; beta.nonzero <- 1; p <- 100
beta <- c(rep(beta.nonzero, p.nonzero), rep(0, p-p.nonzero))
Sigma <- matrix(rep(rho, p*p), p, p); diag(Sigma) <- rep(1, p)
dat <- MASS::mvrnorm(n, rep(0,p), Sigma)
Sigma.eigen <- eigen(Sigma)
mu <- unlist(base::lapply(X=1:p, FUN=function(ii) sqrt(n)*beta[ii]*sqrt(var(dat[, ii]))/sig))
fdp(0.01)
}
}
