% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fdr.r
\name{my.fun}
\alias{my.fun}
\title{A full experient using fdp()}
\usage{
my.fun(p, t)
}
\arguments{
\item{p}{prob}

\item{t}{point to be estimated}
}
\value{
plots
}
\description{
A full experient using fdp()
}
\examples{
\dontrun{
set.seed(12345)
n <- 100; rho <- 0.5; sig <- 2; p.nonzero <- 10; beta.nonzero <- 1
my.fun(p=100, t=0.01)
}
}
