% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_fdr.r
\name{my.plot}
\alias{my.plot}
\title{A full experient using my.fdp()}
\usage{
my.plot(p, t)
}
\arguments{
\item{p}{Number of parameters in the linear model}

\item{t}{point to be estimated}
}
\value{
plots
}
\description{
A full experient using my.fdp()
}
\examples{
\dontrun{
set.seed(12345)
n <- 100; rho <- 0.5; sig <- 2; p.nonzero <- 10; beta.nonzero <- 1
my.plot(p=100, t=0.01)
}
}
