\name{SC19086}
\alias{SC19086}
\docType{package}
\title{
  Statistics Computation Course Final Project
}
\description{
  Statistics Computation Course Final Project
  Include: 4 individual functions with detail information and example and all homeworks with rmd and html.
}
\details{
  This section should provide a more detailed overview of how to use the package, including the most important functions.
  
  1. An FDP estimation method descripted by "Fan, J., Han, X. and Gu, W., 2012. Estimating false discovery proportion under arbitrary covariance dependence. Journal of the American Statistical Association, 107(499), pp.1019-1035."

  2. An KDE method descripted by "Liu, H., Lafferty, J. and Wasserman, L., 2007, March. Sparse nonparametric density estimation in high dimensions using the rodeo. In Artificial Intelligence and Statistics (pp. 283-290)."
}
\author{
Huihang Liu, dawn.gnius@outlook.com.

Maintainer: hhl <dawn.gnius@outlook.com>
}
\references{
  1. Liu, H., Lafferty, J. and Wasserman, L., 2007, March. Sparse nonparametric density estimation in high dimensions using the rodeo. In Artificial Intelligence and Statistics (pp. 283-290)

  2. Fan, J., Han, X. and Gu, W., 2012. Estimating false discovery proportion under arbitrary covariance dependence. Journal of the American Statistical Association, 107(499), pp.1019-1035.
}
\keyword{package, computation, rodeo, fdp}
\seealso{
  https://github.com/DawnGnius/SC19086
}