\name{StatComp19086-package}
\alias{StatComp19086-package}
\alias{StatComp19086}
\docType{package}
\title{
  Statistics Computation Course Final Project
}
\description{
  Statistics Computation Course Final Project
  Include: 3 individual functions with detail information and example and all homeworks with rmd and html.
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Huihang Liu, dawn.gnius@outlook.com.

Maintainer: hhl <dawn.gnius@outlook.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     rw_Metropolis(sigma, x0, N)
  }
}
