% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rw_Metropolis}
\alias{rw_Metropolis}
\title{A Metropolis sampler using Rcpp}
\usage{
rw_Metropolis(sigma, x0, N)
}
\arguments{
\item{sigma}{variance}

\item{x0}{the number of between-sample random number}

\item{N}{the number of samples}
}
\value{
a List with a random sample of size \code{n} and k
}
\description{
A Metropolis sampler using Rcpp
}
\examples{
\dontrun{
sigma <- 2
x0 <- 25
N <- 2000
res <- rw_Metropolis(sigma, x0, N)
}
}
